+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Ability to configure prices for spaces by time slots different than the default hourly rate
 - Updated portuguese translation
 - Refactored the ReserveButton component to use the same user's data across all the component
-- [TODO DEPLOY] `rails db:seed`
+- First optimization the load time of the payment schedules list
 
 ## v5.1.13 2021 November 16